pyproject.toml
[project]
name = "fastapi-tv-series"
version = "0.1.0"
description = "FastAPI TV series catalogue with SQLModel + SQLite"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.122.0",
    "httpx==0.28.1",
    "passlib==1.7.4",
    "pydantic-ai==0.0.19",
    "python-dotenv==1.0.1",
    "pyjwt==2.10.1",
    "redis==5.2.1",
    "requests==2.32.5",
    "sqlmodel==0.0.27",
    "streamlit==1.52.2",
    "uvicorn==0.38.0",
    "typer==0.20.0",
]

[dependency-groups]
dev = [
    "fakeredis==2.31.1",
    "pytest==8.4.1",
    "ruff==0.14.7",
    "schemathesis==3.39.16",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
